Conflicts:
	Sampoerna.EMS.BLL/BLLMapper.cs
	Sampoerna.EMS.BLL/BLLMapperPBCK1.cs
	Sampoerna.EMS.BLL/PBCK1BLL.cs
	Sampoerna.EMS.BLL/WorkflowBLL.cs
	Sampoerna.EMS.Website/Sampoerna.EMS.Website.csproj

+++ Sampoerna.EMS.BusinessObject/DTOs/Pbck1ProdConverterDto.cs
@@ -3,7 +3,7 @@
     public class Pbck1ProdConverterDto
     {
         public long Pbck1ProdConvId { get; set; }
-        public long Pbck1Id { get; set; }
+        public int Pbck1Id { get; set; }
         public string ProdTypeCode { get; set; }
         public string ProdTypeName { get; set; }
         public string ProdAlias { get; set; }
